+++ src/components/service/analytics/doughnutChart/DoughnutChart.module.scss
@@ -1,8 +1,13 @@
+.card {
+    height: 100%;
+    padding: 2.5rem;
+    border-radius: var(--border-radius);
+    background-color: var(--color-2);
+    box-shadow: var(--shadow);
+}
+
 .chart {
     position: relative;
     width: 100%;
     height: 100%;
-    border-radius: var(--border-radius);
-    background-color: var(--color-2);
-    box-shadow: var(--shadow);
 }
